+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+v1.5.1
+v1.5.0
+Error correction and optimization resulted in a 80% reduction in the time to export PDF and IMG files. The ability to export only to PDF without having to generate images was added.
+Fixed create dirs if exportImg and exportPDF are false.
+
 v1.4.14
 Remove pdfkit dependency, using canvas instead. Now this package has only one dependency, canvas.